Version 0.1.5 + backend API sync + NotificationOperator initiation fixed

+++ rtcms4j-spring-client/src/main/kotlin/ru/enzhine/rtcms4j/spring/client/config/props/Rtcms4jProperties.kt
@@ -11,5 +11,5 @@ data class Rtcms4jProperties(
     val applicationId: Long,
     val tokenRefreshOffset: Duration = Duration.ofSeconds(30),
     val clientName: String = UUID.randomUUID().toString(),
-    val pageSize: Long = 20,
+    val pageSize: Int = 20,
 )
